Reduce memory size by 8kb per chunk

+++ Source/World/Chunk/Nodes.h
@@ -47,7 +47,7 @@ struct CBlock
     }
 
     uint16_t id         : 12;  //Maximum value -> 4095
-    uint8_t metaData    : 4;   //Maximum value -> 15
+    uint16_t metaData    : 4;   //Maximum value -> 15
 };
 
 struct CLight
